moved existing migration tests into one file
